Replace placeholder compiled matcher with independent NFA backend
The previous CompiledPattern was a byte-for-byte copy of the oracle's
recursive backtracker, so the oracle/compiled parity property tests were
a tautology that could never catch a semantic bug.

Replace it with a genuinely independent forward NFA-style simulation:
a single left-to-right sweep per partition maintaining in-flight partial
matches ("threads"), each carrying its own bindings (registers), last
consumed index (window/absence anchor), and an incremental absence flag.
The two backends now share only leaf predicate evaluation, so a parity
divergence is a real bug.

This immediately surfaced one: FirstSuccessorPerStart was backtracking to
a later successor when the first dead-ended downstream, contradicting the
spec ("earliest successor satisfying the next step and its transition").
Fix the oracle to commit per step; pin it with a deterministic regression
test, since the proptest generators never produced the divergent shape.

Also fold in review cleanups:
- make ==/!= compare numerically across Int/Float, consistent with the
  ordering operators; compare Int/Int exactly to avoid f64 precision loss
- add From<f64> for Value
- debug_assert input is partition/time sorted at both entry points
  (previously is_sorted_by_partition_time_index was dead code)
- document the tiny parser's non-quote-aware limitations
- doc MatchConsumption variants; clippy tidy (let-chains, derived Default)

cargo fmt, clippy, and 25 tests all pass; parity holds at 8192 proptest
cases.

+++ crates/epigrep-core/benches/matcher.rs
@@ -17,7 +17,7 @@ fn synthetic_events() -> Vec<Event> {
             };
 
             Event::new("partition-1", index, event_type)
-                .with_attr("score", ((index % 5) as i64).into())
+                .with_attr("score", (index % 5).into())
                 .with_attr("user_id", user.into())
         })
         .collect()
